Major change! in led setup
Moved the Led Config of the Forms/Strips to a new config files. So that its posible to load in the patterns without changing the led layout. also added auto calculation for palletes compression. So that its possible to take the same paterns and play them on the diffrenet lamps and they still look good and not squed